+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 History
 -------
 
+**2021-08-04 (0.2.0)**
+ - Add function MenuWithDefault()
+
 **2021-05-12 (0.1.1)**
  - Bugfixes in GetBytes()